+++ src/regs/cntv_tval_el0.rs
@@ -14,9 +14,9 @@ use register::cpu::RegisterReadWrite;
 
 pub struct Reg;
 
-impl RegisterReadWrite<u32, ()> for Reg {
-    sys_coproc_read_raw!(u32, "CNTV_TVAL_EL0");
-    sys_coproc_write_raw!(u32, "CNTV_TVAL_EL0");
+impl RegisterReadWrite<u64, ()> for Reg {
+    sys_coproc_read_raw!(u64, "CNTV_TVAL_EL0");
+    sys_coproc_write_raw!(u64, "CNTV_TVAL_EL0");
 }
 
 pub static CNTV_TVAL_EL0: Reg = Reg {};
