+++ src/regs/spsr_el2.rs
@@ -11,7 +11,7 @@
 
 use register::{cpu::RegisterReadWrite, register_bitfields};
 
-register_bitfields! {u32,
+register_bitfields! {u64,
     pub SPSR_EL2 [
         /// Negative condition flag.
         ///
@@ -129,9 +129,9 @@ register_bitfields! {u32,
 
 pub struct Reg;
 
-impl RegisterReadWrite<u32, SPSR_EL2::Register> for Reg {
-    sys_coproc_read_raw!(u32, "SPSR_EL2");
-    sys_coproc_write_raw!(u32, "SPSR_EL2");
+impl RegisterReadWrite<u64, SPSR_EL2::Register> for Reg {
+    sys_coproc_read_raw!(u64, "SPSR_EL2");
+    sys_coproc_write_raw!(u64, "SPSR_EL2");
 }
 
 pub static SPSR_EL2: Reg = Reg {};
